LayoutTests/imported/w3c/web-platform-tests/css/css-values/calc-size/animation/interpolate-size-max-height-interpolation-expected.txt


PASS CSS Transitions with transition-behavior:allow-discrete: property <max-height> from [initial] to [20px] at (-0.3) should be [initial]
PASS CSS Transitions with transition-behavior:allow-discrete: property <max-height> from [initial] to [20px] at (0) should be [initial]
PASS CSS Transitions with transition-behavior:allow-discrete: property <max-height> from [initial] to [20px] at (0.3) should be [initial]
PASS CSS Transitions with transition-behavior:allow-discrete: property <max-height> from [initial] to [20px] at (0.5) should be [20px]
PASS CSS Transitions with transition-behavior:allow-discrete: property <max-height> from [initial] to [20px] at (0.6) should be [20px]
PASS CSS Transitions with transition-behavior:allow-discrete: property <max-height> from [initial] to [20px] at (1) should be [20px]
PASS CSS Transitions with transition-behavior:allow-discrete: property <max-height> from [initial] to [20px] at (1.5) should be [20px]
PASS CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <max-height> from [initial] to [20px] at (-0.3) should be [initial]
PASS CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <max-height> from [initial] to [20px] at (0) should be [initial]
PASS CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <max-height> from [initial] to [20px] at (0.3) should be [initial]
PASS CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <max-height> from [initial] to [20px] at (0.5) should be [20px]
PASS CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <max-height> from [initial] to [20px] at (0.6) should be [20px]
PASS CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <max-height> from [initial] to [20px] at (1) should be [20px]
PASS CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <max-height> from [initial] to [20px] at (1.5) should be [20px]
PASS CSS Transitions: property <max-height> from [initial] to [20px] at (-0.3) should be [20px]
PASS CSS Transitions: property <max-height> from [initial] to [20px] at (0) should be [20px]
PASS CSS Transitions: property <max-height> from [initial] to [20px] at (0.3) should be [20px]
PASS CSS Transitions: property <max-height> from [initial] to [20px] at (0.5) should be [20px]
PASS CSS Transitions: property <max-height> from [initial] to [20px] at (0.6) should be [20px]
PASS CSS Transitions: property <max-height> from [initial] to [20px] at (1) should be [20px]
PASS CSS Transitions: property <max-height> from [initial] to [20px] at (1.5) should be [20px]
PASS CSS Transitions with transition: all: property <max-height> from [initial] to [20px] at (-0.3) should be [20px]
PASS CSS Transitions with transition: all: property <max-height> from [initial] to [20px] at (0) should be [20px]
PASS CSS Transitions with transition: all: property <max-height> from [initial] to [20px] at (0.3) should be [20px]
PASS CSS Transitions with transition: all: property <max-height> from [initial] to [20px] at (0.5) should be [20px]
PASS CSS Transitions with transition: all: property <max-height> from [initial] to [20px] at (0.6) should be [20px]
PASS CSS Transitions with transition: all: property <max-height> from [initial] to [20px] at (1) should be [20px]
PASS CSS Transitions with transition: all: property <max-height> from [initial] to [20px] at (1.5) should be [20px]
PASS CSS Animations: property <max-height> from [initial] to [20px] at (-0.3) should be [initial]
PASS CSS Animations: property <max-height> from [initial] to [20px] at (0) should be [initial]
PASS CSS Animations: property <max-height> from [initial] to [20px] at (0.3) should be [initial]
PASS CSS Animations: property <max-height> from [initial] to [20px] at (0.5) should be [20px]
PASS CSS Animations: property <max-height> from [initial] to [20px] at (0.6) should be [20px]
PASS CSS Animations: property <max-height> from [initial] to [20px] at (1) should be [20px]
PASS CSS Animations: property <max-height> from [initial] to [20px] at (1.5) should be [20px]
PASS Web Animations: property <max-height> from [initial] to [20px] at (-0.3) should be [initial]
PASS Web Animations: property <max-height> from [initial] to [20px] at (0) should be [initial]
PASS Web Animations: property <max-height> from [initial] to [20px] at (0.3) should be [initial]
PASS Web Animations: property <max-height> from [initial] to [20px] at (0.5) should be [20px]
PASS Web Animations: property <max-height> from [initial] to [20px] at (0.6) should be [20px]
PASS Web Animations: property <max-height> from [initial] to [20px] at (1) should be [20px]
PASS Web Animations: property <max-height> from [initial] to [20px] at (1.5) should be [20px]
PASS CSS Transitions with transition-behavior:allow-discrete: property <max-height> from [unset] to [20px] at (-0.3) should be [unset]
PASS CSS Transitions with transition-behavior:allow-discrete: property <max-height> from [unset] to [20px] at (0) should be [unset]
PASS CSS Transitions with transition-behavior:allow-discrete: property <max-height> from [unset] to [20px] at (0.3) should be [unset]
PASS CSS Transitions with transition-behavior:allow-discrete: property <max-height> from [unset] to [20px] at (0.5) should be [20px]
PASS CSS Transitions with transition-behavior:allow-discrete: property <max-height> from [unset] to [20px] at (0.6) should be [20px]
PASS CSS Transitions with transition-behavior:allow-discrete: property <max-height> from [unset] to [20px] at (1) should be [20px]
PASS CSS Transitions with transition-behavior:allow-discrete: property <max-height> from [unset] to [20px] at (1.5) should be [20px]
PASS CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <max-height> from [unset] to [20px] at (-0.3) should be [unset]
PASS CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <max-height> from [unset] to [20px] at (0) should be [unset]
PASS CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <max-height> from [unset] to [20px] at (0.3) should be [unset]
PASS CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <max-height> from [unset] to [20px] at (0.5) should be [20px]
PASS CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <max-height> from [unset] to [20px] at (0.6) should be [20px]
PASS CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <max-height> from [unset] to [20px] at (1) should be [20px]
PASS CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <max-height> from [unset] to [20px] at (1.5) should be [20px]
PASS CSS Transitions: property <max-height> from [unset] to [20px] at (-0.3) should be [20px]
PASS CSS Transitions: property <max-height> from [unset] to [20px] at (0) should be [20px]
PASS CSS Transitions: property <max-height> from [unset] to [20px] at (0.3) should be [20px]
PASS CSS Transitions: property <max-height> from [unset] to [20px] at (0.5) should be [20px]
PASS CSS Transitions: property <max-height> from [unset] to [20px] at (0.6) should be [20px]
PASS CSS Transitions: property <max-height> from [unset] to [20px] at (1) should be [20px]
PASS CSS Transitions: property <max-height> from [unset] to [20px] at (1.5) should be [20px]
PASS CSS Transitions with transition: all: property <max-height> from [unset] to [20px] at (-0.3) should be [20px]
PASS CSS Transitions with transition: all: property <max-height> from [unset] to [20px] at (0) should be [20px]
PASS CSS Transitions with transition: all: property <max-height> from [unset] to [20px] at (0.3) should be [20px]
PASS CSS Transitions with transition: all: property <max-height> from [unset] to [20px] at (0.5) should be [20px]
PASS CSS Transitions with transition: all: property <max-height> from [unset] to [20px] at (0.6) should be [20px]
PASS CSS Transitions with transition: all: property <max-height> from [unset] to [20px] at (1) should be [20px]
PASS CSS Transitions with transition: all: property <max-height> from [unset] to [20px] at (1.5) should be [20px]
PASS CSS Animations: property <max-height> from [unset] to [20px] at (-0.3) should be [unset]
PASS CSS Animations: property <max-height> from [unset] to [20px] at (0) should be [unset]
PASS CSS Animations: property <max-height> from [unset] to [20px] at (0.3) should be [unset]
PASS CSS Animations: property <max-height> from [unset] to [20px] at (0.5) should be [20px]
PASS CSS Animations: property <max-height> from [unset] to [20px] at (0.6) should be [20px]
PASS CSS Animations: property <max-height> from [unset] to [20px] at (1) should be [20px]
PASS CSS Animations: property <max-height> from [unset] to [20px] at (1.5) should be [20px]
PASS Web Animations: property <max-height> from [unset] to [20px] at (-0.3) should be [unset]
PASS Web Animations: property <max-height> from [unset] to [20px] at (0) should be [unset]
PASS Web Animations: property <max-height> from [unset] to [20px] at (0.3) should be [unset]
PASS Web Animations: property <max-height> from [unset] to [20px] at (0.5) should be [20px]
PASS Web Animations: property <max-height> from [unset] to [20px] at (0.6) should be [20px]
PASS Web Animations: property <max-height> from [unset] to [20px] at (1) should be [20px]
PASS Web Animations: property <max-height> from [unset] to [20px] at (1.5) should be [20px]
PASS CSS Transitions with transition-behavior:allow-discrete: property <max-height> from [none] to [20px] at (-0.3) should be [none]
PASS CSS Transitions with transition-behavior:allow-discrete: property <max-height> from [none] to [20px] at (0) should be [none]
PASS CSS Transitions with transition-behavior:allow-discrete: property <max-height> from [none] to [20px] at (0.3) should be [none]
PASS CSS Transitions with transition-behavior:allow-discrete: property <max-height> from [none] to [20px] at (0.5) should be [20px]
PASS CSS Transitions with transition-behavior:allow-discrete: property <max-height> from [none] to [20px] at (0.6) should be [20px]
PASS CSS Transitions with transition-behavior:allow-discrete: property <max-height> from [none] to [20px] at (1) should be [20px]
PASS CSS Transitions with transition-behavior:allow-discrete: property <max-height> from [none] to [20px] at (1.5) should be [20px]
PASS CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <max-height> from [none] to [20px] at (-0.3) should be [none]
PASS CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <max-height> from [none] to [20px] at (0) should be [none]
PASS CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <max-height> from [none] to [20px] at (0.3) should be [none]
PASS CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <max-height> from [none] to [20px] at (0.5) should be [20px]
PASS CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <max-height> from [none] to [20px] at (0.6) should be [20px]
PASS CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <max-height> from [none] to [20px] at (1) should be [20px]
PASS CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <max-height> from [none] to [20px] at (1.5) should be [20px]
PASS CSS Transitions: property <max-height> from [none] to [20px] at (-0.3) should be [20px]
PASS CSS Transitions: property <max-height> from [none] to [20px] at (0) should be [20px]
PASS CSS Transitions: property <max-height> from [none] to [20px] at (0.3) should be [20px]
PASS CSS Transitions: property <max-height> from [none] to [20px] at (0.5) should be [20px]
PASS CSS Transitions: property <max-height> from [none] to [20px] at (0.6) should be [20px]
PASS CSS Transitions: property <max-height> from [none] to [20px] at (1) should be [20px]
PASS CSS Transitions: property <max-height> from [none] to [20px] at (1.5) should be [20px]
PASS CSS Transitions with transition: all: property <max-height> from [none] to [20px] at (-0.3) should be [20px]
PASS CSS Transitions with transition: all: property <max-height> from [none] to [20px] at (0) should be [20px]
PASS CSS Transitions with transition: all: property <max-height> from [none] to [20px] at (0.3) should be [20px]
PASS CSS Transitions with transition: all: property <max-height> from [none] to [20px] at (0.5) should be [20px]
PASS CSS Transitions with transition: all: property <max-height> from [none] to [20px] at (0.6) should be [20px]
PASS CSS Transitions with transition: all: property <max-height> from [none] to [20px] at (1) should be [20px]
PASS CSS Transitions with transition: all: property <max-height> from [none] to [20px] at (1.5) should be [20px]
PASS CSS Animations: property <max-height> from [none] to [20px] at (-0.3) should be [none]
PASS CSS Animations: property <max-height> from [none] to [20px] at (0) should be [none]
PASS CSS Animations: property <max-height> from [none] to [20px] at (0.3) should be [none]
PASS CSS Animations: property <max-height> from [none] to [20px] at (0.5) should be [20px]
PASS CSS Animations: property <max-height> from [none] to [20px] at (0.6) should be [20px]
PASS CSS Animations: property <max-height> from [none] to [20px] at (1) should be [20px]
PASS CSS Animations: property <max-height> from [none] to [20px] at (1.5) should be [20px]
PASS Web Animations: property <max-height> from [none] to [20px] at (-0.3) should be [none]
PASS Web Animations: property <max-height> from [none] to [20px] at (0) should be [none]
PASS Web Animations: property <max-height> from [none] to [20px] at (0.3) should be [none]
PASS Web Animations: property <max-height> from [none] to [20px] at (0.5) should be [20px]
PASS Web Animations: property <max-height> from [none] to [20px] at (0.6) should be [20px]
PASS Web Animations: property <max-height> from [none] to [20px] at (1) should be [20px]
PASS Web Animations: property <max-height> from [none] to [20px] at (1.5) should be [20px]
PASS CSS Transitions with transition-behavior:allow-discrete: property <max-height> from [none] to [max-content] at (-0.3) should be [none]
PASS CSS Transitions with transition-behavior:allow-discrete: property <max-height> from [none] to [max-content] at (0) should be [none]
PASS CSS Transitions with transition-behavior:allow-discrete: property <max-height> from [none] to [max-content] at (0.3) should be [none]
PASS CSS Transitions with transition-behavior:allow-discrete: property <max-height> from [none] to [max-content] at (0.5) should be [max-content]
PASS CSS Transitions with transition-behavior:allow-discrete: property <max-height> from [none] to [max-content] at (0.6) should be [max-content]
PASS CSS Transitions with transition-behavior:allow-discrete: property <max-height> from [none] to [max-content] at (1) should be [max-content]
PASS CSS Transitions with transition-behavior:allow-discrete: property <max-height> from [none] to [max-content] at (1.5) should be [max-content]
PASS CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <max-height> from [none] to [max-content] at (-0.3) should be [none]
PASS CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <max-height> from [none] to [max-content] at (0) should be [none]
PASS CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <max-height> from [none] to [max-content] at (0.3) should be [none]
PASS CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <max-height> from [none] to [max-content] at (0.5) should be [max-content]
PASS CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <max-height> from [none] to [max-content] at (0.6) should be [max-content]
PASS CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <max-height> from [none] to [max-content] at (1) should be [max-content]
PASS CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <max-height> from [none] to [max-content] at (1.5) should be [max-content]
PASS CSS Transitions: property <max-height> from [none] to [max-content] at (-0.3) should be [max-content]
PASS CSS Transitions: property <max-height> from [none] to [max-content] at (0) should be [max-content]
PASS CSS Transitions: property <max-height> from [none] to [max-content] at (0.3) should be [max-content]
PASS CSS Transitions: property <max-height> from [none] to [max-content] at (0.5) should be [max-content]
PASS CSS Transitions: property <max-height> from [none] to [max-content] at (0.6) should be [max-content]
PASS CSS Transitions: property <max-height> from [none] to [max-content] at (1) should be [max-content]
PASS CSS Transitions: property <max-height> from [none] to [max-content] at (1.5) should be [max-content]
PASS CSS Transitions with transition: all: property <max-height> from [none] to [max-content] at (-0.3) should be [max-content]
PASS CSS Transitions with transition: all: property <max-height> from [none] to [max-content] at (0) should be [max-content]
PASS CSS Transitions with transition: all: property <max-height> from [none] to [max-content] at (0.3) should be [max-content]
PASS CSS Transitions with transition: all: property <max-height> from [none] to [max-content] at (0.5) should be [max-content]
PASS CSS Transitions with transition: all: property <max-height> from [none] to [max-content] at (0.6) should be [max-content]
PASS CSS Transitions with transition: all: property <max-height> from [none] to [max-content] at (1) should be [max-content]
PASS CSS Transitions with transition: all: property <max-height> from [none] to [max-content] at (1.5) should be [max-content]
PASS CSS Animations: property <max-height> from [none] to [max-content] at (-0.3) should be [none]
PASS CSS Animations: property <max-height> from [none] to [max-content] at (0) should be [none]
PASS CSS Animations: property <max-height> from [none] to [max-content] at (0.3) should be [none]
PASS CSS Animations: property <max-height> from [none] to [max-content] at (0.5) should be [max-content]
PASS CSS Animations: property <max-height> from [none] to [max-content] at (0.6) should be [max-content]
PASS CSS Animations: property <max-height> from [none] to [max-content] at (1) should be [max-content]
PASS CSS Animations: property <max-height> from [none] to [max-content] at (1.5) should be [max-content]
PASS Web Animations: property <max-height> from [none] to [max-content] at (-0.3) should be [none]
PASS Web Animations: property <max-height> from [none] to [max-content] at (0) should be [none]
PASS Web Animations: property <max-height> from [none] to [max-content] at (0.3) should be [none]
PASS Web Animations: property <max-height> from [none] to [max-content] at (0.5) should be [max-content]
PASS Web Animations: property <max-height> from [none] to [max-content] at (0.6) should be [max-content]
PASS Web Animations: property <max-height> from [none] to [max-content] at (1) should be [max-content]
PASS Web Animations: property <max-height> from [none] to [max-content] at (1.5) should be [max-content]
FAIL CSS Transitions with transition-behavior:allow-discrete: property <max-height> from [stretch] to [fit-content] at (-0.3) should be [stretch] assert_true: 'from' value should be supported expected true got false
FAIL CSS Transitions with transition-behavior:allow-discrete: property <max-height> from [stretch] to [fit-content] at (0) should be [stretch] assert_true: 'from' value should be supported expected true got false
FAIL CSS Transitions with transition-behavior:allow-discrete: property <max-height> from [stretch] to [fit-content] at (0.3) should be [stretch] assert_true: 'from' value should be supported expected true got false
FAIL CSS Transitions with transition-behavior:allow-discrete: property <max-height> from [stretch] to [fit-content] at (0.5) should be [fit-content] assert_true: 'from' value should be supported expected true got false
FAIL CSS Transitions with transition-behavior:allow-discrete: property <max-height> from [stretch] to [fit-content] at (0.6) should be [fit-content] assert_true: 'from' value should be supported expected true got false
FAIL CSS Transitions with transition-behavior:allow-discrete: property <max-height> from [stretch] to [fit-content] at (1) should be [fit-content] assert_true: 'from' value should be supported expected true got false
FAIL CSS Transitions with transition-behavior:allow-discrete: property <max-height> from [stretch] to [fit-content] at (1.5) should be [fit-content] assert_true: 'from' value should be supported expected true got false
FAIL CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <max-height> from [stretch] to [fit-content] at (-0.3) should be [stretch] assert_true: 'from' value should be supported expected true got false
FAIL CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <max-height> from [stretch] to [fit-content] at (0) should be [stretch] assert_true: 'from' value should be supported expected true got false
FAIL CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <max-height> from [stretch] to [fit-content] at (0.3) should be [stretch] assert_true: 'from' value should be supported expected true got false
FAIL CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <max-height> from [stretch] to [fit-content] at (0.5) should be [fit-content] assert_true: 'from' value should be supported expected true got false
FAIL CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <max-height> from [stretch] to [fit-content] at (0.6) should be [fit-content] assert_true: 'from' value should be supported expected true got false
FAIL CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <max-height> from [stretch] to [fit-content] at (1) should be [fit-content] assert_true: 'from' value should be supported expected true got false
FAIL CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <max-height> from [stretch] to [fit-content] at (1.5) should be [fit-content] assert_true: 'from' value should be supported expected true got false
FAIL CSS Transitions: property <max-height> from [stretch] to [fit-content] at (-0.3) should be [fit-content] assert_true: 'from' value should be supported expected true got false
FAIL CSS Transitions: property <max-height> from [stretch] to [fit-content] at (0) should be [fit-content] assert_true: 'from' value should be supported expected true got false
FAIL CSS Transitions: property <max-height> from [stretch] to [fit-content] at (0.3) should be [fit-content] assert_true: 'from' value should be supported expected true got false
FAIL CSS Transitions: property <max-height> from [stretch] to [fit-content] at (0.5) should be [fit-content] assert_true: 'from' value should be supported expected true got false
FAIL CSS Transitions: property <max-height> from [stretch] to [fit-content] at (0.6) should be [fit-content] assert_true: 'from' value should be supported expected true got false
FAIL CSS Transitions: property <max-height> from [stretch] to [fit-content] at (1) should be [fit-content] assert_true: 'from' value should be supported expected true got false
FAIL CSS Transitions: property <max-height> from [stretch] to [fit-content] at (1.5) should be [fit-content] assert_true: 'from' value should be supported expected true got false
FAIL CSS Transitions with transition: all: property <max-height> from [stretch] to [fit-content] at (-0.3) should be [fit-content] assert_true: 'from' value should be supported expected true got false
FAIL CSS Transitions with transition: all: property <max-height> from [stretch] to [fit-content] at (0) should be [fit-content] assert_true: 'from' value should be supported expected true got false
FAIL CSS Transitions with transition: all: property <max-height> from [stretch] to [fit-content] at (0.3) should be [fit-content] assert_true: 'from' value should be supported expected true got false
FAIL CSS Transitions with transition: all: property <max-height> from [stretch] to [fit-content] at (0.5) should be [fit-content] assert_true: 'from' value should be supported expected true got false
FAIL CSS Transitions with transition: all: property <max-height> from [stretch] to [fit-content] at (0.6) should be [fit-content] assert_true: 'from' value should be supported expected true got false
FAIL CSS Transitions with transition: all: property <max-height> from [stretch] to [fit-content] at (1) should be [fit-content] assert_true: 'from' value should be supported expected true got false
FAIL CSS Transitions with transition: all: property <max-height> from [stretch] to [fit-content] at (1.5) should be [fit-content] assert_true: 'from' value should be supported expected true got false
FAIL CSS Animations: property <max-height> from [stretch] to [fit-content] at (-0.3) should be [stretch] assert_true: 'from' value should be supported expected true got false
FAIL CSS Animations: property <max-height> from [stretch] to [fit-content] at (0) should be [stretch] assert_true: 'from' value should be supported expected true got false
FAIL CSS Animations: property <max-height> from [stretch] to [fit-content] at (0.3) should be [stretch] assert_true: 'from' value should be supported expected true got false
FAIL CSS Animations: property <max-height> from [stretch] to [fit-content] at (0.5) should be [fit-content] assert_true: 'from' value should be supported expected true got false
FAIL CSS Animations: property <max-height> from [stretch] to [fit-content] at (0.6) should be [fit-content] assert_true: 'from' value should be supported expected true got false
FAIL CSS Animations: property <max-height> from [stretch] to [fit-content] at (1) should be [fit-content] assert_true: 'from' value should be supported expected true got false
FAIL CSS Animations: property <max-height> from [stretch] to [fit-content] at (1.5) should be [fit-content] assert_true: 'from' value should be supported expected true got false
FAIL Web Animations: property <max-height> from [stretch] to [fit-content] at (-0.3) should be [stretch] assert_true: 'from' value should be supported expected true got false
FAIL Web Animations: property <max-height> from [stretch] to [fit-content] at (0) should be [stretch] assert_true: 'from' value should be supported expected true got false
FAIL Web Animations: property <max-height> from [stretch] to [fit-content] at (0.3) should be [stretch] assert_true: 'from' value should be supported expected true got false
FAIL Web Animations: property <max-height> from [stretch] to [fit-content] at (0.5) should be [fit-content] assert_true: 'from' value should be supported expected true got false
FAIL Web Animations: property <max-height> from [stretch] to [fit-content] at (0.6) should be [fit-content] assert_true: 'from' value should be supported expected true got false
FAIL Web Animations: property <max-height> from [stretch] to [fit-content] at (1) should be [fit-content] assert_true: 'from' value should be supported expected true got false
FAIL Web Animations: property <max-height> from [stretch] to [fit-content] at (1.5) should be [fit-content] assert_true: 'from' value should be supported expected true got false
FAIL CSS Transitions: property <max-height> from [150px] to [min-content] at (-0.3) should be [calc-size(min-content, 195px + size * -0.3)] assert_equals: expected "100px " but got "min - content "
FAIL CSS Transitions: property <max-height> from [150px] to [min-content] at (0) should be [calc-size(min-content, 150px + size * 0)] assert_equals: expected "100px " but got "min - content "
FAIL CSS Transitions: property <max-height> from [150px] to [min-content] at (0.3) should be [calc-size(min-content, 105px + size * 0.3)] assert_equals: expected "100px " but got "min - content "
FAIL CSS Transitions: property <max-height> from [150px] to [min-content] at (0.6) should be [calc-size(min-content, 60px + size * 0.6)] assert_equals: expected "100px " but got "min - content "
FAIL CSS Transitions: property <max-height> from [150px] to [min-content] at (1) should be [calc-size(min-content, 0px + size * 1)] assert_equals: expected "100px " but got "min - content "
FAIL CSS Transitions: property <max-height> from [150px] to [min-content] at (1.5) should be [calc-size(min-content, -75px + size * 1.5)] assert_equals: expected "100px " but got "min - content "
FAIL CSS Transitions with transition: all: property <max-height> from [150px] to [min-content] at (-0.3) should be [calc-size(min-content, 195px + size * -0.3)] assert_equals: expected "100px " but got "min - content "
FAIL CSS Transitions with transition: all: property <max-height> from [150px] to [min-content] at (0) should be [calc-size(min-content, 150px + size * 0)] assert_equals: expected "100px " but got "min - content "
FAIL CSS Transitions with transition: all: property <max-height> from [150px] to [min-content] at (0.3) should be [calc-size(min-content, 105px + size * 0.3)] assert_equals: expected "100px " but got "min - content "
FAIL CSS Transitions with transition: all: property <max-height> from [150px] to [min-content] at (0.6) should be [calc-size(min-content, 60px + size * 0.6)] assert_equals: expected "100px " but got "min - content "
FAIL CSS Transitions with transition: all: property <max-height> from [150px] to [min-content] at (1) should be [calc-size(min-content, 0px + size * 1)] assert_equals: expected "100px " but got "min - content "
FAIL CSS Transitions with transition: all: property <max-height> from [150px] to [min-content] at (1.5) should be [calc-size(min-content, -75px + size * 1.5)] assert_equals: expected "100px " but got "min - content "
FAIL CSS Animations: property <max-height> from [150px] to [min-content] at (-0.3) should be [calc-size(min-content, 195px + size * -0.3)] assert_equals: expected "100px " but got "150px "
FAIL CSS Animations: property <max-height> from [150px] to [min-content] at (0) should be [calc-size(min-content, 150px + size * 0)] assert_equals: expected "100px " but got "150px "
FAIL CSS Animations: property <max-height> from [150px] to [min-content] at (0.3) should be [calc-size(min-content, 105px + size * 0.3)] assert_equals: expected "100px " but got "150px "
FAIL CSS Animations: property <max-height> from [150px] to [min-content] at (0.6) should be [calc-size(min-content, 60px + size * 0.6)] assert_equals: expected "100px " but got "min - content "
FAIL CSS Animations: property <max-height> from [150px] to [min-content] at (1) should be [calc-size(min-content, 0px + size * 1)] assert_equals: expected "100px " but got "min - content "
FAIL CSS Animations: property <max-height> from [150px] to [min-content] at (1.5) should be [calc-size(min-content, -75px + size * 1.5)] assert_equals: expected "100px " but got "min - content "
FAIL Web Animations: property <max-height> from [150px] to [min-content] at (-0.3) should be [calc-size(min-content, 195px + size * -0.3)] assert_equals: expected "100px " but got "150px "
FAIL Web Animations: property <max-height> from [150px] to [min-content] at (0) should be [calc-size(min-content, 150px + size * 0)] assert_equals: expected "100px " but got "150px "
FAIL Web Animations: property <max-height> from [150px] to [min-content] at (0.3) should be [calc-size(min-content, 105px + size * 0.3)] assert_equals: expected "100px " but got "150px "
FAIL Web Animations: property <max-height> from [150px] to [min-content] at (0.6) should be [calc-size(min-content, 60px + size * 0.6)] assert_equals: expected "100px " but got "min - content "
FAIL Web Animations: property <max-height> from [150px] to [min-content] at (1) should be [calc-size(min-content, 0px + size * 1)] assert_equals: expected "100px " but got "min - content "
FAIL Web Animations: property <max-height> from [150px] to [min-content] at (1.5) should be [calc-size(min-content, -75px + size * 1.5)] assert_equals: expected "100px " but got "min - content "
PASS CSS Transitions: property <max-height> from [min-content] to neutral at (-0.3) should be [calc-size(min-content, -30px + size * 1.3)]
PASS CSS Transitions: property <max-height> from [min-content] to neutral at (0) should be [calc-size(min-content, 0px + size * 1)]
PASS CSS Transitions: property <max-height> from [min-content] to neutral at (0.3) should be [calc-size(min-content, 30px + size * 0.7)]
PASS CSS Transitions: property <max-height> from [min-content] to neutral at (0.6) should be [calc-size(min-content, 60px + size * 0.4)]
PASS CSS Transitions: property <max-height> from [min-content] to neutral at (1) should be [calc-size(min-content, 100px + size * 0)]
PASS CSS Transitions: property <max-height> from [min-content] to neutral at (1.5) should be [calc-size(min-content, 150px + size * -0.5)]
PASS CSS Transitions with transition: all: property <max-height> from [min-content] to neutral at (-0.3) should be [calc-size(min-content, -30px + size * 1.3)]
PASS CSS Transitions with transition: all: property <max-height> from [min-content] to neutral at (0) should be [calc-size(min-content, 0px + size * 1)]
PASS CSS Transitions with transition: all: property <max-height> from [min-content] to neutral at (0.3) should be [calc-size(min-content, 30px + size * 0.7)]
PASS CSS Transitions with transition: all: property <max-height> from [min-content] to neutral at (0.6) should be [calc-size(min-content, 60px + size * 0.4)]
PASS CSS Transitions with transition: all: property <max-height> from [min-content] to neutral at (1) should be [calc-size(min-content, 100px + size * 0)]
PASS CSS Transitions with transition: all: property <max-height> from [min-content] to neutral at (1.5) should be [calc-size(min-content, 150px + size * -0.5)]
FAIL CSS Animations: property <max-height> from [min-content] to neutral at (-0.3) should be [calc-size(min-content, -30px + size * 1.3)] assert_equals: expected "100px " but got "min - content "
FAIL CSS Animations: property <max-height> from [min-content] to neutral at (0) should be [calc-size(min-content, 0px + size * 1)] assert_equals: expected "100px " but got "min - content "
FAIL CSS Animations: property <max-height> from [min-content] to neutral at (0.3) should be [calc-size(min-content, 30px + size * 0.7)] assert_equals: expected "100px " but got "min - content "
PASS CSS Animations: property <max-height> from [min-content] to neutral at (0.6) should be [calc-size(min-content, 60px + size * 0.4)]
PASS CSS Animations: property <max-height> from [min-content] to neutral at (1) should be [calc-size(min-content, 100px + size * 0)]
PASS CSS Animations: property <max-height> from [min-content] to neutral at (1.5) should be [calc-size(min-content, 150px + size * -0.5)]
FAIL Web Animations: property <max-height> from [min-content] to neutral at (-0.3) should be [calc-size(min-content, -30px + size * 1.3)] assert_equals: expected "100px " but got "min - content "
FAIL Web Animations: property <max-height> from [min-content] to neutral at (0) should be [calc-size(min-content, 0px + size * 1)] assert_equals: expected "100px " but got "min - content "
FAIL Web Animations: property <max-height> from [min-content] to neutral at (0.3) should be [calc-size(min-content, 30px + size * 0.7)] assert_equals: expected "100px " but got "min - content "
PASS Web Animations: property <max-height> from [min-content] to neutral at (0.6) should be [calc-size(min-content, 60px + size * 0.4)]
PASS Web Animations: property <max-height> from [min-content] to neutral at (1) should be [calc-size(min-content, 100px + size * 0)]
PASS Web Animations: property <max-height> from [min-content] to neutral at (1.5) should be [calc-size(min-content, 150px + size * -0.5)]
FAIL CSS Transitions: property <max-height> from neutral to [max-content] at (-0.3) should be [calc-size(max-content, 130px + size * -0.3)] assert_equals: expected "100px " but got "max - content "
FAIL CSS Transitions: property <max-height> from neutral to [max-content] at (0) should be [calc-size(max-content, 100px + size * 0)] assert_equals: expected "100px " but got "max - content "
FAIL CSS Transitions: property <max-height> from neutral to [max-content] at (0.3) should be [calc-size(max-content, 70px + size * 0.3)] assert_equals: expected "100px " but got "max - content "
FAIL CSS Transitions: property <max-height> from neutral to [max-content] at (0.6) should be [calc-size(max-content, 40px + size * 0.6)] assert_equals: expected "100px " but got "max - content "
FAIL CSS Transitions: property <max-height> from neutral to [max-content] at (1) should be [calc-size(max-content, 0px + size * 1)] assert_equals: expected "100px " but got "max - content "
FAIL CSS Transitions: property <max-height> from neutral to [max-content] at (1.5) should be [calc-size(max-content, -50px + size * 1.5)] assert_equals: expected "100px " but got "max - content "
FAIL CSS Transitions with transition: all: property <max-height> from neutral to [max-content] at (-0.3) should be [calc-size(max-content, 130px + size * -0.3)] assert_equals: expected "100px " but got "max - content "
FAIL CSS Transitions with transition: all: property <max-height> from neutral to [max-content] at (0) should be [calc-size(max-content, 100px + size * 0)] assert_equals: expected "100px " but got "max - content "
FAIL CSS Transitions with transition: all: property <max-height> from neutral to [max-content] at (0.3) should be [calc-size(max-content, 70px + size * 0.3)] assert_equals: expected "100px " but got "max - content "
FAIL CSS Transitions with transition: all: property <max-height> from neutral to [max-content] at (0.6) should be [calc-size(max-content, 40px + size * 0.6)] assert_equals: expected "100px " but got "max - content "
FAIL CSS Transitions with transition: all: property <max-height> from neutral to [max-content] at (1) should be [calc-size(max-content, 0px + size * 1)] assert_equals: expected "100px " but got "max - content "
FAIL CSS Transitions with transition: all: property <max-height> from neutral to [max-content] at (1.5) should be [calc-size(max-content, -50px + size * 1.5)] assert_equals: expected "100px " but got "max - content "
PASS CSS Animations: property <max-height> from neutral to [max-content] at (-0.3) should be [calc-size(max-content, 130px + size * -0.3)]
PASS CSS Animations: property <max-height> from neutral to [max-content] at (0) should be [calc-size(max-content, 100px + size * 0)]
PASS CSS Animations: property <max-height> from neutral to [max-content] at (0.3) should be [calc-size(max-content, 70px + size * 0.3)]
FAIL CSS Animations: property <max-height> from neutral to [max-content] at (0.6) should be [calc-size(max-content, 40px + size * 0.6)] assert_equals: expected "100px " but got "max - content "
FAIL CSS Animations: property <max-height> from neutral to [max-content] at (1) should be [calc-size(max-content, 0px + size * 1)] assert_equals: expected "100px " but got "max - content "
FAIL CSS Animations: property <max-height> from neutral to [max-content] at (1.5) should be [calc-size(max-content, -50px + size * 1.5)] assert_equals: expected "100px " but got "max - content "
PASS Web Animations: property <max-height> from neutral to [max-content] at (-0.3) should be [calc-size(max-content, 130px + size * -0.3)]
PASS Web Animations: property <max-height> from neutral to [max-content] at (0) should be [calc-size(max-content, 100px + size * 0)]
PASS Web Animations: property <max-height> from neutral to [max-content] at (0.3) should be [calc-size(max-content, 70px + size * 0.3)]
FAIL Web Animations: property <max-height> from neutral to [max-content] at (0.6) should be [calc-size(max-content, 40px + size * 0.6)] assert_equals: expected "100px " but got "max - content "
FAIL Web Animations: property <max-height> from neutral to [max-content] at (1) should be [calc-size(max-content, 0px + size * 1)] assert_equals: expected "100px " but got "max - content "
FAIL Web Animations: property <max-height> from neutral to [max-content] at (1.5) should be [calc-size(max-content, -50px + size * 1.5)] assert_equals: expected "100px " but got "max - content "
